requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
sqlalchemy==2.0.48
psycopg2-binary==2.9.9
psycopg[binary]==3.3.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
alembic==1.13.1
apscheduler==3.10.4
bcrypt==3.2.0
